# Backend/src/Aurigma.DirectMail.Sample.TypeScriptApiClient/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Direct Mail API",
    "version": "v1"
  },
  "paths": {
...
    "/api/direct-mail/v1/integrated-products": {
      "get": {
        "tags": [
          "IntegratedProduct"
        ],
        "summary": "Returns a list of all integrated products.",
        "operationId": "IntegratedProduct_GetIntegratedProducts",
        "parameters": [
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/integrated-products/{id}/update-resources": {
      "post": {
        "tags": [
          "IntegratedProduct"
        ],
        "summary": "Updates a integrated product resources.",
        "operationId": "IntegratedProduct_UpdateIntegratedProductResources",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/integrated-products/{id}/options": {
      "get": {
        "tags": [
          "IntegratedProduct"
        ],
        "summary": "Returns a list of all integrated product options.",
        "operationId": "IntegratedProduct_GetIntegratedProductOptions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductOptionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductOptionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductOptionDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/integrated-products/{id}/templates": {
      "post": {
        "tags": [
          "IntegratedProduct"
        ],
        "summary": "Returns a list of all integrated product templates.",
        "operationId": "IntegratedProduct_GetIntegratedProductTemplates",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Defines options filter.",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationProductOptionRequestModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationProductOptionRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationProductOptionRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/IntegrationProductOptionRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductTemplateDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductTemplateDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IntegratedProductTemplateDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/integrated-products/{id}/templates/{templateId}": {
      "get": {
        "tags": [
          "IntegratedProduct"
        ],
        "summary": "Returns a integrated product template details.",
        "operationId": "IntegratedProduct_GetIntegratedProductTemplateDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "templateId",
            "in": "path",
            "description": "Template id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "ProductVariantId",
            "in": "query",
            "description": "Customer's Canvas product id.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/IntegratedProductTemplateDetailsDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegratedProductTemplateDetailsDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegratedProductTemplateDetailsDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConflictDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/direct-mail/v1/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Returns a list of all products.",
        "operationId": "Product_GetProducts",
        "parameters": [
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "Creates a new product.",
        "operationId": "Product_CreateProduct",
        "requestBody": {
          "description": "Creation model Aurigma.DirectMail.Sample.WebApi.Models.Products.ProductCreationModel",
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreationModel"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreationModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreationModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreationModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/direct-mail/v1/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "Returns a product by id.",
        "operationId": "Product_GetProduct",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product's id.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConflictDto": {
        "type": "object",
        "properties": {
          "propertyName": {
            "type": "string",
            "nullable": true
          },
          "propertyValue": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "IntegratedProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier.",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "description": "Product title.",
            "nullable": true
          },
          "price": {
            "type": "number",
            "description": "Product price.",
            "format": "double"
          },
          "previewUrl": {
            "type": "string",
            "description": "Customer's canvas preview url.",
            "nullable": true
          },
          "customersCanvasProductId": {
            "type": "integer",
            "description": "Product ID in Customers Canvas.",
            "format": "int32"
          },
          "categoryId": {
            "type": "string",
            "description": "Product category id.",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IntegratedProductOptionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "optionType": {
            "$ref": "#/components/schemas/IntegratedProductOptionType"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntegratedProductOptionValueDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IntegratedProductOptionType": {
        "enum": [
          "Radio",
          "Dropdown",
          "Chips",
          "ColorGrid",
          "ColorList",
          "ImageGrid"
        ],
        "type": "string"
      },
      "IntegratedProductOptionValueDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "sortIndex": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "IntegratedProductTemplateDetailsDto": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "string",
            "nullable": true
          },
          "templateName": {
            "type": "string",
            "nullable": true
          },
          "productVariantId": {
            "type": "integer",
            "format": "int32"
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntegratedProductTemplateDetailsOptionDto"
            },
            "nullable": true
          },
          "previewUrls": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "customFields": {
            "type": "object",
            "additionalProperties": { },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IntegratedProductTemplateDetailsOptionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IntegratedProductTemplateDto": {
        "type": "object",
        "properties": {
          "templateId": {
            "type": "string",
            "nullable": true
          },
          "templateName": {
            "type": "string",
            "nullable": true
          },
          "productVariantId": {
            "type": "integer",
            "format": "int32"
          },
          "previewUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "IntegrationProductOptionItemRequestModel": {
        "type": "object",
        "properties": {
          "optionId": {
            "type": "integer",
            "description": "Customer's canvas product option id.",
            "format": "int32"
          },
          "valueIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "description": "/ Customer's canvas product option value ids.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "IntegrationProductOptionRequestModel": {
        "type": "object",
        "properties": {
          "templateTitle": {
            "type": "string",
            "description": "",
            "nullable": true
          },
          "options": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IntegrationProductOptionItemRequestModel"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "ProductCreationModel": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}